代码优化
1. 重构了所有插件，Close #353
2. 重写了权限机制，Close #355
