+++ README.md
@@ -177,18 +177,18 @@ npm run restart
 
 | 功能 | 形式 | 权限开关 | 主人命令 |
 | --- | --- | --- | --- |
-| 展示米游社 ID 、 UID 或者某个群友的游戏账号 | 插件 | ✔️ | ❌ |
-| 展示米游社 ID 、 UID 或者某个群友的深渊战绩 | 插件 | ✔️ | ❌ |
-| 米游社ID绑定和改绑 | 插件 | ❌ | ❌ |
+| 展示米游社、 UID 或者某个群友的游戏账号 | 插件 | ✔️ | ❌ |
+| 展示米游社、 UID 或者某个群友的深渊战绩 | 插件 | ✔️ | ❌ |
+| 米游社ID绑定和改绑 | 插件 | ✔️ | ❌ |
 | 圣遗物掉落和强化 | 插件 | ✔️ | ❌ |
 | 圣遗物截图评分 | 插件 | ✔️ | ❌ |
 | 展示角色官方数据 | 插件 | ✔️ | ❌ |
 | 祈愿十连（支持定轨） | 插件 | ✔️ | ❌ |
-| 今天该刷什么 | 插件 | ❌ | ❌ |
-| 今天吃什么 | 插件 | ❌ | ❌ |
+| 今天该刷什么 | 插件 | ✔️ | ❌ |
+| 今天吃什么 | 插件 | ✔️ | ❌ |
 | 点歌 | 插件 | ✔️ | ❌ |
-| 掷骰子 | 插件 | ❌ | ❌ |
-| 求签 | 插件 | ❌ | ❌ |
+| 掷骰子 | 插件 | ✔️ | ❌ |
+| 求签 | 插件 | ✔️ | ❌ |
 | 主人和其他好友或群聊天、发送广播 | 插件 | ❌ | ✔️ |
 | 查看、搜索和统计添加的好友和群 | 插件 | ❌ | ✔️ |
 | 群广播和好友广播 | 插件 | ❌ | ✔️ |
