+++ src/plugins/tools/index.js
@@ -1,6 +1,7 @@
 /* global command, master */
 /* eslint no-undef: "error" */
 
+import { checkAuth } from "../../utils/auth.js";
 import { hasEntrance } from "../../utils/config.js";
 import { feedback } from "./feedback.js";
 import { menu } from "./menu.js";
@@ -8,30 +9,38 @@ import { prophecy } from "./prophecy.js";
 import { quote } from "./quote.js";
 import { roll } from "./roll.js";
 
-async function Plugin(msg, bot) {
-  const groupName = "group" === msg.type ? msg.group_name : undefined;
-
+async function Plugin(msg) {
   switch (true) {
     case hasEntrance(msg.text, "tools", "menu"):
-      menu(msg.sid, msg.text, msg.type, msg.uid, bot);
+      if (false !== (await checkAuth(msg, "menu"))) {
+        await menu(msg);
+      }
       break;
     case hasEntrance(msg.text, "tools", "prophecy"):
-      prophecy(msg.sid, msg.text, msg.type, msg.uid, bot);
+      if (false !== (await checkAuth(msg, "prophecy"))) {
+        await prophecy(msg);
+      }
       break;
     case hasEntrance(msg.text, "tools", "roll"):
-      roll(msg.sid, msg.text, msg.type, msg.uid, bot);
+      if (false !== (await checkAuth(msg, "roll"))) {
+        await roll(msg);
+      }
       break;
     case hasEntrance(msg.text, "tools", "quote"):
-      quote(msg.sid, msg.text, msg.type, msg.uid, bot);
+      if (false !== (await checkAuth(msg, "quote"))) {
+        await quote(msg);
+      }
       break;
     case hasEntrance(msg.text, "tools", "feedback"):
-      feedback(msg.sid, msg.name, msg.text, msg.type, msg.uid, groupName, bot);
+      if (false !== (await checkAuth(msg, "feedback"))) {
+        await feedback(msg);
+      }
       break;
     case hasEntrance(msg.text, "tools", "help"):
-      await bot.say(msg.sid, command.usage, msg.type, msg.uid, "\n");
+      await msg.bot.say(msg.sid, command.usage, msg.type, msg.uid, "\n");
       break;
     case hasEntrance(msg.text, "tools", "master"):
-      await bot.say(msg.sid, master.usage, msg.type, msg.uid, "\n");
+      await msg.bot.say(msg.sid, master.usage, msg.type, msg.uid, "\n");
       break;
   }
 }
