+++ src/plugins/tools/menu.js
@@ -5,14 +5,14 @@ import { getRandomInt } from "../../utils/tools.js";
 
 const { breakfast, lunch, dinner } = config.menu;
 
-async function menu(id, text, type, user, bot) {
-  const favFood = "派蒙";
+async function menu(msg) {
+  const food = "派蒙";
   const message = `今日的推荐菜单是：
-早餐：${breakfast ? breakfast[getRandomInt(breakfast.length) - 1] : favFood}
-午餐：${lunch ? lunch[getRandomInt(lunch.length) - 1] : favFood}
-晚餐：${dinner ? dinner[getRandomInt(dinner.length) - 1] : favFood}`;
+早餐：${breakfast ? breakfast[getRandomInt(breakfast.length) - 1] : food}
+午餐：${lunch ? lunch[getRandomInt(lunch.length) - 1] : food}
+晚餐：${dinner ? dinner[getRandomInt(dinner.length) - 1] : food}`;
 
-  await bot.say(id, message, type, user);
+  await msg.bot.say(msg.sid, message, msg.type, msg.uid);
 }
 
 export { menu };
