+++ src/plugins/tools/roll.js
@@ -1,7 +1,7 @@
 import { getRandomInt } from "../../utils/tools.js";
 
-async function roll(id, text, type, user, bot) {
-  let cmd = text.match(/[+-]?[0-9]+/g);
+async function roll(msg) {
+  let cmd = msg.text.match(/[+-]?[0-9]+/g);
 
   if (null === cmd) {
     cmd = ["100"];
@@ -10,7 +10,7 @@ async function roll(id, text, type, user, bot) {
   if (1 === cmd.length) {
     const max = parseInt(cmd[0]);
     const res = max < 1 || max > 100 ? "骰子面数应为不超过 100 的正整数。" : `骰子的结果为: ${getRandomInt(max)}。`;
-    await bot.say(id, res, type, user);
+    await msg.bot.say(msg.sid, res, msg.type, msg.uid);
   }
 }
