+++ src/plugins/tools_master/count.js
@@ -1,11 +1,11 @@
-async function count(id, text, type, user, bot) {
+async function count(msg) {
   let report = "";
   let num = 0;
 
-  report += `好友个数：${bot.fl.size - 1}`; // 排除好友列表中的自己
-  report += `\n群组个数：${bot.gl.size}`;
+  report += `好友个数：${msg.bot.fl.size - 1}`; // 排除好友列表中的自己
+  report += `\n群组个数：${msg.bot.gl.size}`;
 
-  bot.gl.forEach((item) => {
+  msg.bot.gl.forEach((item) => {
     if (item) {
       num += item.member_count;
       num--; // 排除群友列表中的自己
@@ -14,7 +14,7 @@ async function count(id, text, type, user, bot) {
 
   report += `\n群友个数：${num}`;
 
-  await bot.say(id, report, type, user, "\n");
+  await msg.bot.say(msg.sid, report, msg.type, msg.uid, "\n");
   return;
 }
