+++ src/plugins/tools_master/reply.js
@@ -1,7 +1,7 @@
-async function reply(id, text, type, user, bot) {
-  const target = parseInt(text.match(/[0-9]+/g)[0]);
-  const [msg] = text.split(/(?<=\d+\S+)\s/).slice(1);
-  const list = new Map([...bot.fl, ...bot.gl]);
+async function reply(msg) {
+  const target = parseInt(msg.text.match(/[0-9]+/g)[0]);
+  const [text] = msg.text.split(/(?<=\d+\S+)\s/).slice(1);
+  const list = new Map([...msg.bot.fl, ...msg.bot.gl]);
 
   list.forEach(async (item) => {
     const curType = item.group_id ? "group" : "private";
@@ -9,9 +9,9 @@ async function reply(id, text, type, user, bot) {
 
     if (itemID == target) {
       // 送话无法 @
-      await bot.say(itemID, `主人让我送个话：\n${msg}`, curType);
+      await msg.bot.say(itemID, `主人让我送个话：\n${text}`, curType);
       // 私聊无法 @
-      await bot.say(id, `我已经给${itemID}送话了。`, "private");
+      await msg.bot.say(msg.sid, `我已经给${itemID}送话了。`, "private");
     }
   });
 }
