+++ src/plugins/tools_master/status.js
@@ -4,7 +4,7 @@ import pb from "pretty-bytes";
 import path from "path";
 import { du } from "../../utils/file.js";
 
-async function status(id, type, user, bot) {
+async function status(msg) {
   const os = await si.osInfo();
   const cpu = await si.cpu();
   const mem = await si.mem();
@@ -19,7 +19,7 @@ CPU：${load.currentLoad && load.currentLoad.toFixed(2)}%（${cpu.manufacturer} 
 启动：${moment.duration(time.uptime * 1000).humanize()}
 数据：${pb(du(path.join("data", "db")))}`;
 
-  await bot.say(id, str, type, user, "\n");
+  await msg.bot.say(msg.sid, str, msg.type, msg.uid, "\n");
 }
 
 export { status };
